+++ src/document_state.rs
@@ -1,5 +1,5 @@
 use derive_new::new;
-use lsp_types::{Hover, HoverParams, Position, Url};
+use lsp_types::{DocumentSymbolResponse, Hover, HoverParams, Position, Url};
 
 use crate::parol_ls_grammar::ParolLsGrammar;
 
@@ -14,12 +14,19 @@ impl DocumentState {
         self.parsed_data.ident_at_position(position)
     }
 
+    pub(crate) fn clear(&mut self) {
+        self.parsed_data = ParolLsGrammar::default()
+    }
+
     pub(crate) fn hover(&self, params: HoverParams) -> Hover {
         self.parsed_data.hover(params, &self.input)
     }
 
-    pub(crate) fn clear(&mut self) {
-        self.parsed_data = ParolLsGrammar::default()
+    pub(crate) fn document_symbols(
+        &self,
+        params: lsp_types::DocumentSymbolParams,
+    ) -> DocumentSymbolResponse {
+        self.parsed_data.document_symbols(params, &self.input)
     }
 }