+++ src/handler.rs
@@ -12,13 +12,11 @@ impl RequestHandler for GotoDefinition {
         eprintln!("got gotoDefinition request #{}: {:?}", id, params);
         let result = server.handle_goto_definition(params);
         let result = serde_json::to_value(&result).unwrap();
-        let resp = Response {
+        Response {
             id,
             result: Some(result),
             error: None,
-        };
-        eprintln!("got gotoDefinition response {:?}", resp);
-        resp
+        }
     }
 }
 
@@ -27,13 +25,11 @@ impl RequestHandler for HoverRequest {
         eprintln!("got hover request #{}: {:?}", id, params);
         let result = server.handle_hover(params);
         let result = serde_json::to_value(&result).unwrap();
-        let resp = Response {
+        Response {
             id,
             result: Some(result),
             error: None,
-        };
-        eprintln!("got hover response {:?}", resp);
-        resp
+        }
     }
 }
 
@@ -42,12 +38,10 @@ impl RequestHandler for DocumentSymbolRequest {
         eprintln!("got document symbols request #{}: {:?}", id, params);
         let result = server.handle_document_symbols(params);
         let result = serde_json::to_value(&result).unwrap();
-        let resp = Response {
+        Response {
             id,
             result: Some(result),
             error: None,
-        };
-        eprintln!("got document symbols response {:?}", resp);
-        resp
+        }
     }
 }